move finalize to base multitemper
