+++ pycbc/inference/sampler/ptemcee.py
@@ -312,7 +312,7 @@ class PTEmceeSampler(MultiTemperedAutocorrSupport, MultiTemperedSupport,
     def calculate_logevidence(cls, filename, thin_start=None, thin_end=None,
                               thin_interval=None):
         """Calculates the log evidence from the given file.
-        
+
         This uese ``ptemcee``'s thermodynamic integration.
 
         Parameters
@@ -367,28 +367,6 @@ class PTEmceeSampler(MultiTemperedAutocorrSupport, MultiTemperedSupport,
                     unique_betas.append(ubti[ii])
         return ptemcee.util.thermodynamic_integration_log_evidence(
             numpy.array(unique_betas), numpy.array(mean_logls))
-
-    def finalize(self):
-        """Calculates the log evidence and writes to the checkpoint file.
-
-        The thin start/interval/end for calculating the log evidence are
-        retrieved from the checkpoint file's thinning attributes.
-        """
-        logging.info("Calculating log evidence")
-        # get the thinning settings
-        with self.io(self.checkpoint_file, 'r') as fp:
-            thin_start = fp.thin_start
-            thin_interval = fp.thin_interval
-            thin_end = fp.thin_end
-        # calculate
-        logz, dlogz = self.calculate_logevidence(
-            self.checkpoint_file, thin_start=thin_start, thin_end=thin_end,
-            thin_interval=thin_interval)
-        logging.info("log Z, dlog Z: {}, {}".format(logz, dlogz))
-        # write to both the checkpoint and backup
-        for fn in [self.checkpoint_file, self.backup_file]:
-            with self.io(fn, "a") as fp:
-                fp.write_logevidence(logz, dlogz)
 
     @classmethod
     def from_config(cls, cp, model, nprocesses=1, use_mpi=False):
